Fix gnutls cert generation failing. Thx owine.
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@7479 e03df62e-2008-0410-955e-edbf42e46eb7

+++ make/gnutlscert.pm
@@ -106,9 +106,10 @@ ocsp_signing_key
 time_stamping_key
 __END__
 close(FH);
-system("certtool --generate-privkey --outfile key.pem") or return 1;
-system("certtool --generate-self-signed --load-privkey key.pem --outfile cert.pem --template certtool.template") or return 1;
+if ( (my $status = system("certtool --generate-privkey --outfile key.pem")) ne 0) { return 1; }
+if ( (my $status = system("certtool --generate-self-signed --load-privkey key.pem --outfile cert.pem --template certtool.template")) ne 0) { return 1; }
 unlink("certtool.template");
+return 0;
 }
 
 1;
